refactor(ui): split menus into package with shared helpers
Replace the monolithic ui/menus.py with a ui/menus/ package split by
flow, plus _common, _display, and _deps for DRY and test patching.
