+++ ui/menus/settings.py
@@ -1,0 +1,119 @@
+"""Настройки, языки и выбор сложности."""
+
+from typing import Any
+
+from colorama import Fore, Style
+
+from core.localization import get_string
+from ui.menus import _deps
+from ui.menus._common import _press_enter, _print_screen_header
+
+
+def select_difficulty(strings: dict[str, Any]) -> str | None:
+    """Экран выбора сложности при создании персонажа."""
+    _print_screen_header(get_string(strings, "difficulty.caption"))
+
+    options = [
+        get_string(strings, "difficulty.normal"),
+        get_string(strings, "difficulty.hardcore"),
+    ]
+    for idx, opt in enumerate(options, 1):
+        marker = f"{Fore.GREEN}* {Style.RESET_ALL}" if idx == 1 else "  "
+        print(f"{marker}{Fore.YELLOW}{idx}{Style.RESET_ALL}. {opt}")
+    print()
+    print(
+        f"  {Fore.YELLOW}0{Style.RESET_ALL}."
+        f" {get_string(strings, 'difficulty.back')}"
+    )
+    print()
+
+    choice = _deps.get_int_input(
+        get_string(strings, "difficulty.prompt", count=len(options)),
+        0,
+        len(options),
+        strings,
+    )
+
+    if choice == 0:
+        return None
+
+    return "normal" if choice == 1 else "hardcore"
+
+
+def show_languages_menu(
+    strings: dict[str, Any], settings: dict[str, Any]
+) -> dict[str, Any]:
+    """Меню выбора языка."""
+    while True:
+        _print_screen_header(get_string(strings, "languages.caption"))
+
+        current = settings.get("language", "ru")
+        lang_name = get_string(
+            strings, f"languages.lang_{current}", default=current
+        )
+        print(f"  {get_string(strings, 'languages.current')} {lang_name}")
+        print()
+
+        lang_codes = ["en", "ru"] if current == "ru" else ["ru", "en"]
+        options = [
+            get_string(strings, f"languages.lang_{code}")
+            for code in lang_codes
+        ]
+        for idx, opt in enumerate(options, 1):
+            marker = f"  {Fore.YELLOW}{idx}{Style.RESET_ALL}. "
+            print(f"{marker}{opt}")
+        print()
+        print(
+            f"  {Fore.YELLOW}0{Style.RESET_ALL}."
+            f" {get_string(strings, 'languages.back')}"
+        )
+        print()
+
+        choice = _deps.get_int_input(
+            get_string(strings, "languages.prompt", count=len(options)),
+            0,
+            len(options),
+            strings,
+        )
+
+        if choice == 0:
+            break
+
+        new_lang = lang_codes[choice - 1]
+        settings["language"] = new_lang
+        strings = _deps.load_strings(new_lang)
+        msg = get_string(
+            strings,
+            "languages.changed",
+            name=get_string(strings, f"languages.lang_{new_lang}"),
+        )
+        print(f"{Fore.GREEN}{msg}{Style.RESET_ALL}")
+        print()
+        _press_enter(strings)
+
+    return settings
+
+
+def show_settings(
+    strings: dict[str, Any], settings: dict[str, Any]
+) -> dict[str, Any]:
+    """Экран настроек."""
+    while True:
+        _print_screen_header(get_string(strings, "settings.caption"))
+        print(
+            f"  {Fore.YELLOW}0{Style.RESET_ALL}."
+            f" {get_string(strings, 'settings.back')}"
+        )
+        print()
+
+        choice = _deps.get_int_input(
+            get_string(strings, "settings.prompt", count=0),
+            0,
+            0,
+            strings,
+        )
+
+        if choice == 0:
+            break
+
+    return settings
